[Android] Reverted size_t back to int

+++ android/java/org/chromium/chrome/browser/onboarding/SearchEngineOnboardingFragment.java
@@ -137,7 +137,7 @@ public class SearchEngineOnboardingFragment extends Fragment {
             @Override
             public void onCheckedChanged(RadioGroup radioGroup, int i) {
                 View radioButton = radioGroup.findViewById(i);
-                size_t index = radioGroup.indexOfChild(radioButton);
+                int index = radioGroup.indexOfChild(radioButton);
                 searchEngineSelected(index, templateUrls);
             }
         });
